backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
google-genai==1.17.0
python-dotenv==1.0.1
gunicorn==23.0.0
google-cloud-firestore
python-docx
structlog
filetype
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
pydantic==2.10.5
xhtml2pdf==0.2.16
